+++ CHANGELOG.md
@@ -8,12 +8,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - changelog file to start log changes (#72)
+- basic unit test for App, ProductsList, ShippingMethods components
+- proper unit test dependencies (vue/test-utils, jest) and configuration
 
 ### Changed
-- 
+- currency vue filter
+- presets in babelrc
 
 ### Removed
-- 
+-
 
 
 ## [0.2.0] - 2018-05-21